+++ Assets/Prefabs/GameManager.prefab
@@ -26,7 +26,7 @@ Transform:
   m_GameObject: {fileID: 4209708903738261937}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 2.292243, y: -0.03958993, z: 0.042274844}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
