+++ Assets/Prefabs/InputManager.prefab
@@ -26,7 +26,7 @@ Transform:
   m_GameObject: {fileID: 5934277780893907704}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0.37821245, y: -0.12342032, z: 0.04010451}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
