+++ Assets/Scripts/CurveController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2deb33634efb24767a4c350ec8ce90c4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
