+++ Assets/Scripts/MarkerSelection.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: acec4b61655a748fab51a59663490eac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
